requirements.txt
Pillow
keras
tensorflow
nltk
numpy
git+https://github.com/farizrahman4u/recurrentshop.git
